refactor(web): split chat rendering components (#419)
