+++ progress.md
@@ -157,7 +157,7 @@
 
 | Change | 状态 | 说明 |
 |---|---|---|
-| entropy-governance | #418 Web Chat model/database gating hook split complete, issues #408-#424 | `.entropy-baseline/latest.json` 已记录实际项目代码熵基线并排除 `external/*`；#410/#411/#412 API error helper 迁移完成；#413-#416 API Chat backend 已拆为 bounded collaborators；#417 Web Chat session/scope/settings hook extraction 完成；#418 将 model availability 与 database gating 拆入 focused hooks，Chat targeted 44 tests、Web 197 tests、typecheck/lint、OpenSpec strict validate 和 diff check 通过 |
+| entropy-governance | #419 Web Chat rendering extraction complete, issues #408-#424 | `.entropy-baseline/latest.json` 已记录实际项目代码熵基线并排除 `external/*`；#410/#411/#412 API error helper 迁移完成；#413-#416 API Chat backend 已拆为 bounded collaborators；#417 Web Chat session/scope/settings hook extraction 完成；#418 model/database gating hooks 完成；#419 将 assistant markdown、message parts、citation panel、source-chain utils/rendering 拆入 `apps/web/src/pages/chat/**`，Chat targeted 48 tests、Web 201 tests、typecheck/lint、OpenSpec strict validate 和 diff check 通过 |
 | graph-explorer-visual-overhaul | panel-theme-alignment complete, issues #400/#401/#403 | 新增 `useGraphTheme` 读取 theme CSS vars 并监听 `data-theme`；GraphCanvas 背景/边框/标签主题化；节点渲染增加默认/选中 glow、选中双环、大图默认 glow 降级；Graph Explorer 面板、Community 选中态和 Legend 对齐 theme token；`getNodeColor`/`getLinkColor` 已提取为纯函数并补充测试 |
 | wiki-version-diff | API+UI complete, pending browser verification | `GET /wiki/pages/:pageId/diff` + version history compare modal；`npm run build`、Wiki API tests 通过 |
 | fix-session-delete-cascade | complete, issue #381 | Chat session 关联 retrieval traces/model usage logs/feedback items 删除级联修复 |
